feat(auth): implement user authentication and session management

+++ src/main/java/com/savemark/app/repositories/UserRepo.java
@@ -3,6 +3,6 @@ package com.savemark.app.repositories;
 import com.savemark.app.models.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<User, Integer> {
+public interface UserRepo extends JpaRepository<User, Long> {
        User findByUsername(String username);
 }
